# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Cinema Platform API",
    "version": "1.0.0",
    "description": "Phase 1. Admin bearer tokens require issuer/audience validation and server-enforced role/cinema scope. Times are UTC; input date filters are inclusive from and exclusive to."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "requestId"
        ]
      },
...
      "FeedbackConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ratingType": {
            "type": "string",
            "enum": [
              "ICON+TEXT",
              "STAR",
              "BUTTON",
              "TEXT"
            ]
          },
          "ratingOptions": {
            "type": "array",
            "minItems": 5,
            "maxItems": 5,
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          },
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reason"
            }
          },
          "minimumFeedbackForRanking": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          },
          "consentVersion": {
            "type": "string"
          }
        },
        "required": [
          "ratingType",
          "ratingOptions",
          "minimumFeedbackForRanking",
          "consentVersion"
        ]
      },
...
      "CinemaPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cinema"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          }
        },
        "required": []
      },
...
    }
  },
  "paths": {
    "/public/feedback/{qrToken}": {
      "get": {
        "summary": "Validate active QR",
        "description": "",
        "operationId": "get_public_feedback_qrToken",
        "security": [],
        "parameters": [
          {
            "name": "qrToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "serverTime": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/public/feedback-config": {
      "get": {
        "summary": "Read customer configuration",
        "description": "",
        "operationId": "get_public_feedback_config",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackConfig"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/public/feedback": {
      "post": {
        "summary": "Submit a new feedback record",
        "description": "No deduplication. Same phone/staff within ten minutes is flagged suspicious. Public API rate limited; Retry-After is returned on HTTP 429.",
        "operationId": "post_public_feedback",
        "security": [],
        "parameters": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submission"
              }
            }
          }
        }
      }
    },
    "/admin/me": {
      "get": {
        "summary": "Current verified identity",
        "description": "",
        "operationId": "get_admin_me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cinemas": {
      "get": {
        "summary": "List cinemas",
        "description": "",
        "operationId": "get_admin_cinemas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "One-based, defaults to 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "1–100, defaults to 20"
          },
          {
            "name": "cinemaId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CinemaPage"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Cinema",
        "description": "",
        "operationId": "post_admin_cinemas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cinema"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cinema"
              }
            }
          }
        }
      }
    },
    "/admin/cinemas/{id}": {
      "put": {
        "summary": "Update Cinema",
        "description": "",
        "operationId": "put_admin_cinemas_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cinema"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cinema"
              }
            }
          }
        }
      }
    },
    "/admin/staff": {
      "get": {
        "summary": "List staff",
        "description": "",
        "operationId": "get_admin_staff",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "One-based, defaults to 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "1–100, defaults to 20"
          },
          {
            "name": "cinemaId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPage"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Staff",
        "description": "",
        "operationId": "post_admin_staff",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            }
          }
        }
      }
    },
    "/admin/staff/{id}": {
      "put": {
        "summary": "Update Staff",
        "description": "",
        "operationId": "put_admin_staff_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Staff"
              }
            }
          }
        }
      }
    },
    "/admin/coaching": {
      "get": {
        "summary": "List coaching",
        "description": "",
        "operationId": "get_admin_coaching",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "One-based, defaults to 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "1–100, defaults to 20"
          },
          {
            "name": "cinemaId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "staffId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "followUp",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoachingPage"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Coaching",
        "description": "",
        "operationId": "post_admin_coaching",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coaching"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coaching"
              }
            }
          }
        }
      }
    },
    "/admin/coaching/{id}": {
      "put": {
        "summary": "Update Coaching",
        "description": "",
        "operationId": "put_admin_coaching_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coaching"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coaching"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Read coaching",
        "description": "",
        "operationId": "get_admin_coaching_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coaching"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "summary": "List users",
        "description": "",
        "operationId": "get_admin_users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "One-based, defaults to 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "1–100, defaults to 20"
          },
          {
            "name": "cinemaId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create User",
        "description": "",
        "operationId": "post_admin_users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
    "/admin/users/{id}": {
      "put": {
        "summary": "Update User",
        "description": "",
        "operationId": "put_admin_users_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
...
    "/admin/feedbacks/{id}": {
      "get": {
        "summary": "Read scoped feedback detail",
        "description": "",
        "operationId": "get_admin_feedbacks_id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feedback"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/audit-logs": {
      "get": {
        "summary": "List scoped audit",
        "description": "",
        "operationId": "get_admin_audit_logs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "One-based, defaults to 1"
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "1–100, defaults to 20"
          },
          {
            "name": "cinemaId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional global filter; managers may only select their own cinema"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Inclusive RFC3339 timestamp"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Exclusive RFC3339 timestamp"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            }
          },
          "default": {
            "description": "Validation, authorization, not found, conflict, rate limit or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}